tag filtering added

+++ src/components/notes-layout.component.tsx
@@ -1,4 +1,4 @@
-import { Box, Button, IconButton } from "@mui/material";
+import { Box, Button, Checkbox, IconButton, ToggleButton } from "@mui/material";
 import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteItem, editItem } from "../app/notes/store/notes.actions";
@@ -6,7 +6,7 @@ import { NotesState } from "../app/notes/types/note-state.type";
 import { TagState } from "../app/tags/types/tag-state.type";
 
 export default function NotesLayout() {
-    const [activeTag, setActiveTag] = useState([''])
+    const [activeTag, setActiveTag] = useState<string[]>([])
 
     function NotesTags() {
         const tags = useSelector((state: TagState) => state.tags);
@@ -15,10 +15,16 @@ export default function NotesLayout() {
             <>
                 {tags.map(tag => (
                     <Box key={tag.text}>
-                        <Button onClick={() => {
-                            setActiveTag([...activeTag, tag.text])
-                            console.log(activeTag)
-                        }}>{tag.text}</Button>
+                        <Button
+                            key={tag.id}
+                            onClick={() => {
+                                if (activeTag.includes(tag.text.toLowerCase())) {
+                                    setActiveTag(activeTag.filter((e) => e !== tag.text))
+                                } else {
+                                    setActiveTag([...activeTag, tag.text])
+                                }
+                            }}>{tag.text}
+                        </Button>
                     </Box>
                 ))}
 
@@ -31,45 +37,47 @@ export default function NotesLayout() {
         const dispatch = useDispatch();
         const note = useSelector((state: NotesState) => state.notes);
         console.log(note)
+        return (<>
+            {
+                note.map(e => {
+                    if (activeTag.every((tag) => e.text.toLowerCase().includes(tag))) {
+                        return (
+                            <Box key={e.id}
+                            >
 
-        return (
-            <>
-                {note.map(e => (
-                    // if ( activeTag.every((tag) => e.text.includes(tag))){}
-                    <Box key={e.id}
-                    >
+                                <input readOnly={readOnlyValue} value={e.text}
+                                    onChange={(i) => {
+                                        dispatch(editItem({
+                                            id: e.id,
+                                            text: i.target.value
+                                        }));
+                                    }}
+                                    onKeyDown={event => {
+                                        if (event.key === 'Enter') {
+                                            setReadOnlyValue(true)
 
-                        <input readOnly={readOnlyValue} value={e.text}
-                            onChange={(i) => {
-                                dispatch(editItem({
-                                    id: e.id,
-                                    text: i.target.value
-                                }));
-                            }}
-                            onKeyDown={event => {
-                                if (event.key === 'Enter') {
-                                    setReadOnlyValue(true)
-
-                                }
-                            }}
-                        />
-                        <IconButton
-                            onClick={() => {
-                                setReadOnlyValue(!readOnlyValue)
-                            }}
-                        >
-                            <span>Edit</span>
-                        </IconButton>
-                        <IconButton
-                            onClick={() => {
-                                dispatch(deleteItem(e));
-                            }}
-                        >
-                            <span>Delete</span>
-                        </IconButton>
-                    </Box>
-                ))}
-            </>
+                                        }
+                                    }}
+                                />
+                                <IconButton
+                                    onClick={() => {
+                                        setReadOnlyValue(!readOnlyValue)
+                                    }}
+                                >
+                                    <span>Edit</span>
+                                </IconButton>
+                                <IconButton
+                                    onClick={() => {
+                                        dispatch(deleteItem(e));
+                                    }}
+                                >
+                                    <span>Delete</span>
+                                </IconButton>
+                            </Box>
+                        )
+                    }
+                })
+            }</>
         )
     }
